% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit.R
\name{BuildNumbers}
\alias{BuildNumbers}
\title{Build the numbers for the tests}
\usage{
BuildNumbers(userlist, geneset, background)
}
\arguments{
\item{userlist}{Vector of user-provided genes}

\item{geneset}{Vector of genes in the gene set}

\item{background}{Vector of all genes in the universe or a number of genes
in the universe}
}
\value{
A list with the following elements:
\itemize{
\item k: Number of genes in the user list that are also in the gene set
\item M: Number of genes in the gene set
\item n: Number of genes in the user list
\item N: Total number of genes in the universe
}
}
\description{
Build the numbers for the tests
}
\keyword{internal}
