% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit.R
\name{ComputePValue}
\alias{ComputePValue}
\title{Compute statistic p-value by given numbers}
\usage{
ComputePValue(k, M, n, N, method = c("fisher", "hypergeometric"))
}
\arguments{
\item{k}{Number of genes in both user list and gene set}

\item{M}{Number of genes in the gene set}

\item{n}{Number of genes in the user list}

\item{N}{Total number of genes in the universe}

\item{method}{Method for computing p-value, either "fisher" or "hypergeometric"}
}
\value{
p-value from the specified test
}
\description{
Compute statistic p-value by given numbers
}
\keyword{internal}
