% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit-package.R
\docType{package}
\name{enrichit-package}
\alias{enrichit}
\alias{enrichit-package}
\title{enrichit: 'C++' Implementations of Functional Enrichment Analysis}
\description{
Fast implementations of functional enrichment analysis methods using 'C++' via 'Rcpp'. Currently provides Over-Representation Analysis (ORA), Gene Set Enrichment Analysis (GSEA), Weighted Enrichment Analysis for ORA and GSEA, Network-based Set Enrichment Analysis (NSEA), multi-layer network-based enrichment, and multi-omics integration workflows. Additional features include early fusion at the feature level, late fusion at the pathway level, multi-omics contribution tracing, topology-aware explanation helpers, Bayesian term selection, and extremely fast Random Walk with Restart (RWR) using 'RcppEigen'. The enrichment methods build on GSEA by Subramanian et al. (2005) \doi{10.1073/pnas.0506580102}, the multilevel strategy derived from 'fgsea' by Korotkevich et al. (2021) \doi{10.1101/060012}, and network-based enrichment ideas described by Glaab et al. (2012) \doi{10.1093/bioinformatics/bts389}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com}
}

}
\keyword{internal}
