% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit.R
\name{EnrichIt}
\alias{EnrichIt}
\title{Perform enrichment analysis}
\usage{
EnrichIt(
  userlist,
  dbs,
  method = ifelse(tolower(style[1]) == "enrichr", "fisher", "hypergeometric"),
  use_matched_only = ifelse(tolower(style[1]) == "enrichr", FALSE, TRUE),
  padjust_method = c("BH", "bonferroni", "holm", "hochberg", "hommel", "BY", "fdr"),
  background = NULL,
  style = c("enrichr", "Enrichr", "clusterProfiler", "ClusterProfiler",
    "clusterprofiler"),
  return_all = FALSE
)
}
\arguments{
\item{userlist}{Vector of user-provided genes}

\item{dbs}{List of gene sets or paths to GMT files
It can be a vector of database names. You can set the names of the vector, which
will be used as the database names.
Otherwise a list is expected, where each element is a database (e.g. parsed from a gmt file).
If a database is given directly (without a name), the expression of this argument will
be used as the database name.}

\item{method}{Method for computing p-value, either "fisher" or "hypergeometric"
When \code{style} is "enrichr", the method defaults to "fisher".
When \code{style} is "clusterProfiler", the method defaults to "hypergeometric".}

\item{use_matched_only}{Logical, whether to use only matched genes against the gene sets.
This will affect the number of genes in the user list when computing the
p-value.
By default, when \code{style} is "enrichr", this is set to FALSE.
When \code{style} is "clusterProfiler", this is set to TRUE.}

\item{padjust_method}{Method for adjusting p-values, either "BH", "bonferroni",
"holm", "hochberg", "hommel", "BY", "fdr"}

\item{background}{Vector of all genes in the universe or a number of genes
in the universe. If NULL, the number of genes in the gene set will be used.
For "enrichr", the default is 20,000.
For "clusterProfiler", the default is the number of unique genes in the gene set.
Note that for "enrichr", if a vector is given, the length of it will be used, no
checking will be done to see if userlist and genes from dbs are in the vector.}

\item{style}{Style of the output, either "enrichr" or "clusterProfiler"}

\item{return_all}{Logical, whether to return all results (all gene sets in dbs)
or only those with at least one gene in the user list.}
}
\value{
A data frame with the results.
When \code{style} is "enrichr", the columns are:
\itemize{
\item Database: Name of the database
\item Term: Name of the term
\item Overlap: Number of genes in the user list and the gene set
\item P.value: p-value from the enrichment test
\item Adjusted.P.value: Adjusted p-value from the enrichment test
\item Odds.Ratio: Odds ratio from the enrichment test
\item Combined.Score: Combined score from the enrichment test
\item Genes: Genes in the user list that are also in the gene set
\item Rank: Rank of the term based on the combined score
}

When \code{style} is "clusterProfiler", the columns are:
\itemize{
\item ID: Name of the term
\item Description: Description of the term
\item GeneRatio: Ratio of genes in the user list and the gene set
\item BgRatio: Ratio of genes in the gene set and the universe
\item Count: Number of genes in the user list that are also in the gene set
\item pvalue: p-value from the enrichment test
\item p.adjust: Adjusted p-value from the enrichment test
\item qvalue: Q-value from the enrichment test
\item geneID: Genes in the user list that are also in the gene set
\item Database: Name of the database
}
}
\description{
Perform enrichment analysis
}
