% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReadable.R
\name{EXTID2NAME}
\alias{EXTID2NAME}
\title{EXTID2NAME}
\usage{
EXTID2NAME(OrgDb, geneID, keytype, toType = "SYMBOL")
}
\arguments{
\item{OrgDb}{OrgDb}

\item{geneID}{entrez gene ID}

\item{keytype}{keytype}

\item{toType}{ID type of the output}
}
\value{
gene symbol
}
\description{
mapping gene ID to gene Symbol
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
