% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{FetchGMT}
\alias{FetchGMT}
\title{Fetch GMT file}
\usage{
FetchGMT(gmtpath)
}
\arguments{
\item{gmtpath}{Path to the GMT file
Either a local file path or a URL.
If a URL is provided, the file will be downloaded and cached.
If a local file path is provided, it will be checked for existence.
It can also be a name of a library from the Enrichr database.
See https://maayanlab.cloud/Enrichr/#libraries.}
}
\value{
If \code{gmtpath} is a local file, it returns the path.
If \code{gmtpath} is a URL, it downloads the file to a temporary cache
directory if not already present and returns the path to the cached file.
}
\description{
Fetch GMT file
}
\examples{
\donttest{
if (FALSE) {
# Fetch a GMT file from a URL (will be cached)
gmtpath_url <- paste0(
  "https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=",
  "Data_Acquisition_Method_Most_Popular_Genes"
)
FetchGMT(gmtpath_url)
# Example output: ~/.cache/enrichit/xxxxxx.gmt

FetchGMT("Data_Acquisition_Method_Most_Popular_Genes")
# Example output: ~/.cache/enrichit/xxxxxx.gmt

# Fetching the same URL again uses the cache
gmt_file_url_cached <- FetchGMT(gmtpath_url)
# Should be the same path as gmt_file_url
print(identical(gmt_file_url, gmt_file_url_cached)) # TRUE

# If the GMT file is local
# Create a dummy local file for example
local_gmt_path <- tempfile(fileext = ".gmt")
writeLines(c("GENESET1\tdesc1\tGENE1\tGENE2", "GENESET2\tdesc2\tGENE3\tGENE4"), local_gmt_path)
FetchGMT(local_gmt_path)
# Example output: /tmp/RtmpXXXXXX/filexxxx.gmt (original path)
}
}
}
\keyword{internal}
