% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneInCategory}
\alias{geneInCategory}
\title{geneInCategory generic}
\usage{
geneInCategory(x)
}
\arguments{
\item{x}{enrichResult}
}
\value{
'geneInCategory' return a list of genes, by spliting the input gene vector to enriched functional categories
}
\description{
geneInCategory generic
}
\examples{
```r
data(geneList, package="DOSE")
de <- names(geneList)[1:100]
x <- DOSE::enrichDO(de)
geneInCategory(x)
```
}
