% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_omics.R
\name{select_features_for_ora}
\alias{select_features_for_ora}
\title{Select features for ORA}
\usage{
select_features_for_ora(x, cutoff = 0.05, by = c("pvalue", "score"), ...)
}
\arguments{
\item{x}{A structured result from \code{aggregate_omics()} or \code{harmonize_ids()}.}

\item{cutoff}{Numeric, the threshold to apply.}

\item{by}{Character, metric to apply the threshold on. One of "pvalue" or "score".}

\item{...}{Additional arguments.}
}
\value{
A list containing \code{gene} (the selected feature IDs) and \code{universe} (all feature IDs).
}
\description{
Convert continuous aggregated statistics into a discrete list of genes and a universe for Over-Representation Analysis.
}
