% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{st_effects_cpp}
\alias{st_effects_cpp}
\title{Effects matrix from .lmEffects (contrast coordinate + residual coordinates)}
\usage{
st_effects_cpp(E, weights, design, contrast_idx)
}
\description{
Effects matrix from .lmEffects (contrast coordinate + residual coordinates)
}
\keyword{internal}
