% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{st_fry_cpp}
\alias{st_fry_cpp}
\title{Native fry: fast rotation-free set test on the effects matrix}
\usage{
st_fry_cpp(E, weights, design, contrast_idx, index, gene_weights, standardize)
}
\description{
Mirrors \code{limma::fry} (standardize = "posterior.sd" by default).
}
\keyword{internal}
