% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{st_squeeze_var_cpp}
\alias{st_squeeze_var_cpp}
\title{limma-style squeezeVar for scalar residual df (fitFDist prior)}
\usage{
st_squeeze_var_cpp(var, df)
}
\description{
limma-style squeezeVar for scalar residual df (fitFDist prior)
}
\keyword{internal}
