% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit-package.R
\docType{package}
\name{enrichit-package}
\alias{enrichit}
\alias{enrichit-package}
\title{enrichit: Fast Over-Representation Analysis (ORA) using C++}
\description{
A fast implementation of Over-Representation Analysis (ORA) using C++ via Rcpp. Provides hypergeometric test (Fisher's exact test) and multiple testing correction.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

}
\keyword{internal}
