% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit-package.R
\docType{package}
\name{enrichit-package}
\alias{enrichit}
\alias{enrichit-package}
\title{enrichit: C++ Implementations of Functional Enrichment Analysis}
\description{
Fast implementations of functional enrichment analysis methods using C++ via Rcpp. Currently provides Over-Representation Analysis (ORA) and will support additional algorithms such as Gene Set Enrichment Analysis (GSEA) in the future.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

}
\keyword{internal}
