% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea_gson}
\alias{gsea_gson}
\title{gsea_gson}
\usage{
gsea_gson(
  geneList,
  gson,
  nPerm = 1000,
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  method = "sample",
  verbose = TRUE,
  seed = FALSE
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{gson}{GSON object}

\item{nPerm}{Number of permutations for p-value calculation (default: 1000).}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{method}{Permutation method: "sample" (default) or "permute"}

\item{verbose}{print message or not}

\item{seed}{set seed inside the function to make result reproducible. FALSE by default.}
}
\value{
gseaResult object
}
\description{
generic function for gene set enrichment analysis
}
\author{
Guangchuang Yu
}
