% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsea.R
\name{nsea_gson}
\alias{nsea_gson}
\title{Network-based GSEA using a GSON object}
\usage{
nsea_gson(
  geneList,
  network,
  gson,
  mode = c("evidence", "signed"),
  p = 0.5,
  minGSSize = 10,
  maxGSSize = 500,
  threshold = 1e-09,
  maxIter = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{geneList}{named numeric vector. In "evidence" mode, must be non-negative. In "signed" mode, can contain both positive and negative values.}

\item{network}{edge list (data.frame) or sparse matrix.}

\item{gson}{a GSON object.}

\item{mode}{character, either "evidence" (default) or "signed".}

\item{p}{restart probability for RWR (default is 0.5).}

\item{minGSSize}{minimal size of each gene set for analyzing. default here is 10.}

\item{maxGSSize}{maximal size of genes annotated for testing. default here is 500.}

\item{threshold}{convergence threshold for RWR (default is 1e-9).}

\item{maxIter}{maximal number of RWR iterations (default is 100).}

\item{verbose}{logical, print messages.}

\item{...}{other arguments passed to \code{gsea_gson()}.}
}
\value{
A \code{gseaResult} object.
}
\description{
Network-based GSEA using a GSON object
}
