% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{nseaResult-class}
\alias{nseaResult-class}
\title{Class "nseaResult"
This class represents the result of Network-based Set Enrichment Analysis (NSEA).}
\description{
Class "nseaResult"
This class represents the result of Network-based Set Enrichment Analysis (NSEA).
}
\section{Slots}{

\describe{
\item{\code{result}}{enrichment analysis}

\item{\code{organism}}{organism label for the enrichment result}

\item{\code{setType}}{gene set collection type}

\item{\code{geneSets}}{gene sets}

\item{\code{geneList}}{order rank geneList}

\item{\code{keytype}}{ID type of gene}

\item{\code{permScores}}{permutation score matrix inherited from \code{gseaResult}}

\item{\code{gene2Symbol}}{gene ID to symbol mapping}

\item{\code{readable}}{logical flag of gene ID in symbol or not.}

\item{\code{termsim}}{Calculation matrix of termsim.}

\item{\code{method}}{Method of termsim.}

\item{\code{params}}{parameters}

\item{\code{dr}}{dimension reduction result}

\item{\code{network}}{sparse matrix or data.frame representing the underlying network.}

\item{\code{diffusion_scores}}{numeric vector of RWR diffusion scores for each node.}

\item{\code{mode}}{character, "evidence" or "signed", describing the RWR propagation mode.}

\item{\code{iterations}}{integer, the actual number of iterations RWR took to converge.}

\item{\code{restart_prob}}{numeric, the restart probability used in RWR.}
}}

\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
