% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{nseaResult-class}
\alias{nseaResult-class}
\alias{show,nseaResult-method}
\alias{summary,nseaResult-method}
\title{Class "nseaResult"
This class represents the result of Network-based Set Enrichment Analysis (NSEA).
It inherits from gseaResult and contains additional network-related attributes.}
\description{
Class "nseaResult"
This class represents the result of Network-based Set Enrichment Analysis (NSEA).
It inherits from gseaResult and contains additional network-related attributes.
}
\section{Slots}{

\describe{
\item{\code{network}}{The input network used for propagation}

\item{\code{diffusion_scores}}{The stationary probabilities (or score differences) after RWR}

\item{\code{mode}}{The propagation mode used ("evidence" or "signed")}

\item{\code{iterations}}{The number of RWR iterations to converge}

\item{\code{restart_prob}}{The restart probability used in RWR}
}}

\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
\keyword{classes}
