% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{ParseGMT}
\alias{ParseGMT}
\title{Read GMT file}
\usage{
ParseGMT(gmtfile, swap_name_desc_if_needed = TRUE)
}
\arguments{
\item{gmtfile}{Path to the GMT file}

\item{swap_name_desc_if_needed}{Logical, whether to swap name and description fields
if all descriptions are shorter than names.}
}
\value{
A list of gene sets, the names of the list are the gene set names,
and each element is a vector of gene names.
}
\description{
Read GMT file
}
\examples{
\donttest{
if (FALSE) {
# Example GMT file content
gmtfile <- system.file("extdata", "KEGG_2021_Human.gmt", package = "enrichit")
# Read the GMT file
ParseGMT(gmtfile)
}
}
}
