% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsea.R
\name{prepare_network}
\alias{prepare_network}
\title{Prepare network for repeated NSEA runs}
\usage{
prepare_network(network, directed = FALSE, normalize = "column")
}
\arguments{
\item{network}{edge list (data.frame with 2 or 3 columns) or sparse matrix.}

\item{directed}{logical, whether the network is directed. Default is FALSE.}

\item{normalize}{one of "column", "row", or "none". Default is "column".}
}
\value{
A sparse matrix (dgCMatrix) that has been properly formatted and normalized.
}
\description{
Prepare network for repeated NSEA runs
}
