% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,nseaResult-method}
\title{show method}
\usage{
show(object)

show(object)

show(object)
}
\arguments{
\item{object}{A \code{enrichResult} instance.}
}
\value{
message

message

message
}
\description{
show method for \code{gseaResult} instance

show method for \code{nseaResult} instance

show method for \code{enrichResult} instance
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
